feat: support custom dependency track via Lagoon api env vars
